+++ src/modules/Experience/Experience.selector.test.ts
@@ -1,9 +1,37 @@
 import { rootStateFixture } from '../../redux/redux.test.fixtures'
+import { USER_JOBS_MOCK } from '../UserJobs/UserJobs.test.fixtures'
 import * as SUT from './Experience.selector'
 
 describe('modules/Experience/Experience.selector', () => {
+  describe('selectUserJobItems', () => {
+    it('should return userJobsItems from userJobs state', () => {
+      const stateMock = rootStateFixture({
+        userJobs: {
+          ids: ['11111-5717-4562-b3fc-2c963f66afa6'],
+          entities: { '11111-5717-4562-b3fc-2c963f66afa6': USER_JOBS_MOCK[0] },
+        },
+      })
+      // when ... we call the selector
+      // @ts-ignore
+      const result = SUT.selectUserJobItems(stateMock)
+
+      // then ... should return result as expected
+      expect(result).toEqual({
+        ids: ['11111-5717-4562-b3fc-2c963f66afa6'],
+        entities: {
+          '11111-5717-4562-b3fc-2c963f66afa6': {
+            title: 'TITLE',
+            subtitle: 'NAME',
+            iconUrl: 'LOGO',
+            isValidated: true,
+            description: 'DESCRIPTION',
+          },
+        },
+      })
+    })
+  })
   describe('default selector', () => {
-    it('should return an empty array by default', () => {
+    it('should handle an empty state', () => {
       const stateMock = rootStateFixture()
       // when ... we call the selector
       // @ts-ignore
@@ -11,7 +39,10 @@ describe('modules/Experience/Experience.selector', () => {
 
       // then ... should return result as expected
       expect(result).toEqual({
-        userJobs: [],
+        userJobs: {
+          ids: [],
+          entities: {},
+        },
         skills: [],
         organisations: [],
       })
@@ -20,7 +51,17 @@ describe('modules/Experience/Experience.selector', () => {
       const mockState = rootStateFixture({
         userJobs: {
           ids: ['id1'],
-          entities: { id1: { job: 'Job Data', startDate: 'Start Data', endDate: 'End Date' } },
+          entities: {
+            id1: {
+              approved: true,
+              job: {
+                title: 'TITLE',
+                description: 'DESCRIPTION',
+                organisationName: 'NAME',
+                organisationLogoURL: 'LOGO',
+              },
+            },
+          },
         },
         skills: {
           ids: ['key1'],
@@ -36,7 +77,18 @@ describe('modules/Experience/Experience.selector', () => {
 
       // then ... should return the data required by the experience view
       expect(result).toEqual({
-        userJobs: [{ job: 'Job Data', startDate: 'Start Data', endDate: 'End Date' }],
+        userJobs: {
+          ids: ['id1'],
+          entities: {
+            id1: {
+              title: 'TITLE',
+              subtitle: 'NAME',
+              iconUrl: 'LOGO',
+              isValidated: true,
+              description: 'DESCRIPTION',
+            },
+          },
+        },
         organisations: [{ label: 'organisation', value: 'key1' }],
         skills: [{ label: 'skill', value: 'skill' }],
       })
